Fixed screen space mapping for model space centered in screen.

+++ src/model/modelparameters.hpp
@@ -18,7 +18,7 @@ struct ModelParameters {
 	//displacement of model in viewport space
 	int displacementX = 0;
 	int displacementY = 0;
-	double zoomLevel = 1.0f;
+	int zoomLevel = 1;
 };
 
 #endif
